refactor: 抽取 _BaseHttpClient 收敛 Boss/招聘者 client 的请求重试与生命周期
