+++ pyproject.toml
@@ -165,6 +165,7 @@ module = [
 	"boss_agent_cli.auth.manager",
 	"boss_agent_cli.auth.qr_login",
 	"boss_agent_cli.api.client",
+	"boss_agent_cli.api._base_client",
 	"boss_agent_cli.api.browser_client",
 	"boss_agent_cli.auth.browser",
 	"boss_agent_cli.bridge.client",
