+++ src/boss_agent_cli/api/client.py
@@ -1,26 +1,9 @@
 import atexit
-import random
-import time
 import weakref
-from types import TracebackType
-from typing import TYPE_CHECKING, Any, cast
-
-import httpx
+from typing import Any
 
 from boss_agent_cli.api import endpoints
-from boss_agent_cli.api.httpx_helpers import (
-	add_stoken_to_get_params,
-	browser_headers,
-	merge_response_cookies,
-	referer_header,
-)
-from boss_agent_cli.api.throttle import RequestThrottle
-
-if TYPE_CHECKING:
-	from boss_agent_cli.api.browser_client import BrowserSession
-	from boss_agent_cli.auth.manager import AuthManager
-
-_MAX_RETRIES = 3
+from boss_agent_cli.api._base_client import _BaseHttpClient
 
 # atexit safeguard: close any BossClient instances not explicitly closed
 _OPEN_CLIENTS: weakref.WeakSet["BossClient"] = weakref.WeakSet()
@@ -49,102 +32,21 @@ class AccountRiskError(Exception):
 		super().__init__(message)
 
 
-class BossClient:
+class BossClient(_BaseHttpClient):
 	"""Hybrid API client: browser channel for high-risk ops, httpx for low-risk ops."""
 
-	def __init__(self, auth_manager: "AuthManager", *, delay: tuple[float, float] = (1.5, 3.0), cdp_url: str | None = None) -> None:
-		self._auth = auth_manager
-		self._delay = delay
-		self._client: httpx.Client | None = None
-		self._browser_session: "BrowserSession | None" = None
-		self._throttle = RequestThrottle(delay)
-		self._cdp_url = cdp_url
-		self._closed = False
+	_BASE_URL = endpoints.BASE_URL
+	_DEFAULT_HEADERS = endpoints.DEFAULT_HEADERS
+	_REFERER_MAP = endpoints.REFERER_MAP
+	_AUTH_ERROR_CLS = AuthError
+	_CODE_STOKEN_EXPIRED = endpoints.CODE_STOKEN_EXPIRED
+	_CODE_RATE_LIMITED = endpoints.CODE_RATE_LIMITED
+
+	def _register(self) -> None:
 		_OPEN_CLIENTS.add(self)
 
-	def _get_client(self) -> httpx.Client:
-		if self._client is None:
-			token = self._auth.get_token()
-			headers = browser_headers(endpoints.DEFAULT_HEADERS, token)
-			self._client = httpx.Client(
-				base_url=endpoints.BASE_URL,
-				cookies=token.get("cookies", {}),
-				headers=headers,
-				follow_redirects=True,
-				timeout=30,
-			)
-		return self._client
-
-	def _get_browser(self) -> "BrowserSession":
-		if self._browser_session is None:
-			from boss_agent_cli.api.browser_client import BrowserSession
-			token = self._auth.get_token()
-			self._browser_session = BrowserSession(
-				cookies=token.get("cookies", {}),
-				user_agent=token.get("user_agent", ""),
-				delay=self._delay,
-				cdp_url=self._cdp_url,
-				logger=getattr(self._auth, '_logger', None),
-			)
-		return self._browser_session
-
-	# ── Anti-detection delays (httpx channel) ────────────────────────
-
-	def _headers_for(self, url: str) -> dict[str, str]:
-		return referer_header(url, endpoints.REFERER_MAP, f"{endpoints.BASE_URL}/")
-
-	def _merge_cookies(self, resp: httpx.Response) -> None:
-		merge_response_cookies(self._get_client(), resp)
-
-	# ── httpx request (low-risk ops) ─────────────────────────────────
-
-	def _request(self, method: str, url: str, **kwargs: Any) -> dict[str, Any]:
-		"""httpx 请求，循环重试（最多 _MAX_RETRIES 次），替代递归调用。"""
-		for attempt in range(_MAX_RETRIES + 1):
-			client = self._get_client()
-			token = self._auth.get_token()
-			stoken = token.get("stoken", "")
-
-			add_stoken_to_get_params(method, kwargs, stoken)
-
-			self._throttle.wait()
-
-			extra_headers = self._headers_for(url)
-			resp = client.request(method, url, headers=extra_headers, **kwargs)
-			self._throttle.mark()
-			self._merge_cookies(resp)
-
-			# 403 或安全验证 → 刷新 token 重试
-			if resp.status_code == 403 or "安全验证" in resp.text:
-				if attempt >= _MAX_RETRIES:
-					raise AuthError("Token 刷新后仍被拒绝，请重新登录")
-				backoff = (2 ** attempt) + random.uniform(0.5, 1.5)
-				time.sleep(backoff)
-				self._auth.force_refresh(cdp_url=self._cdp_url)
-				self._client = None
-				continue
-
-			resp.raise_for_status()
-			data = resp.json()
-			code = data.get("code")
-
-			# stoken 过期 → 刷新重试
-			if code == endpoints.CODE_STOKEN_EXPIRED and attempt < _MAX_RETRIES:
-				backoff = (2 ** attempt) + random.uniform(0.5, 1.5)
-				time.sleep(backoff)
-				self._auth.force_refresh(cdp_url=self._cdp_url)
-				self._client = None
-				continue
-
-			# 频率限制 → 冷却重试
-			if code == endpoints.CODE_RATE_LIMITED and attempt < _MAX_RETRIES:
-				cooldown = min(60, 10 * (2 ** attempt))
-				time.sleep(cooldown)
-				continue
-
-			return cast("dict[str, Any]", data)
-
-		raise AuthError("请求失败，已达最大重试次数")
+	def _unregister(self) -> None:
+		_OPEN_CLIENTS.discard(self)
 
 	# ── Browser request (high-risk ops) ──────────────────────────────
 
@@ -297,29 +199,3 @@ class BossClient:
 		"""查询与某招聘者的互动关系（是否已打招呼等）。"""
 		params = {"securityId": security_id}
 		return self._request("GET", endpoints.GEEK_GET_JOB_URL, params=params)
-
-	# ── Lifecycle ────────────────────────────────────────────────────
-
-	def close(self) -> None:
-		"""Release httpx client and browser session. Idempotent."""
-		if self._closed:
-			return
-		self._closed = True
-		if self._browser_session:
-			self._browser_session.close()
-			self._browser_session = None
-		if self._client:
-			self._client.close()
-			self._client = None
-		_OPEN_CLIENTS.discard(self)
-
-	def __enter__(self) -> "BossClient":
-		return self
-
-	def __exit__(
-		self,
-		exc_type: type[BaseException] | None,
-		exc_val: BaseException | None,
-		exc_tb: TracebackType | None,
-	) -> None:
-		self.close()
