+++ tests/test_recruiter_client_retry.py
@@ -97,9 +97,9 @@ def test_request_extra_headers_override_is_merged():
 	assert http_client.calls[0]["headers"]["Referer"] == "https://override"
 
 
-@patch("boss_agent_cli.api.recruiter_client.random.uniform", return_value=0)
-@patch("boss_agent_cli.api.recruiter_client.time.sleep")
-@patch("boss_agent_cli.api.recruiter_client.httpx.Client")
+@patch("boss_agent_cli.api._base_client.random.uniform", return_value=0)
+@patch("boss_agent_cli.api._base_client.time.sleep")
+@patch("boss_agent_cli.api._base_client.httpx.Client")
 def test_request_retries_after_403_and_refreshes_token(mock_http_client_cls, mock_sleep, mock_uniform):
 	auth = FakeAuthManager()
 	first = FakeHttpxClient([FakeResponse(status_code=403, text="forbidden")])
@@ -118,9 +118,9 @@ def test_request_retries_after_403_and_refreshes_token(mock_http_client_cls, moc
 	assert second.calls[0]["kwargs"]["params"]["__zp_stoken__"] == "refreshed-1"
 
 
-@patch("boss_agent_cli.api.recruiter_client.random.uniform", return_value=0)
-@patch("boss_agent_cli.api.recruiter_client.time.sleep")
-@patch("boss_agent_cli.api.recruiter_client.httpx.Client")
+@patch("boss_agent_cli.api._base_client.random.uniform", return_value=0)
+@patch("boss_agent_cli.api._base_client.time.sleep")
+@patch("boss_agent_cli.api._base_client.httpx.Client")
 def test_request_retries_after_stoken_expired_code(mock_http_client_cls, mock_sleep, mock_uniform):
 	auth = FakeAuthManager()
 	first = FakeHttpxClient([FakeResponse(payload={"code": ep.CODE_STOKEN_EXPIRED})])
@@ -138,8 +138,8 @@ def test_request_retries_after_stoken_expired_code(mock_http_client_cls, mock_sl
 	assert mock_sleep.call_args_list[0].args[0] == 1
 
 
-@patch("boss_agent_cli.api.recruiter_client.time.sleep")
-@patch("boss_agent_cli.api.recruiter_client.httpx.Client")
+@patch("boss_agent_cli.api._base_client.time.sleep")
+@patch("boss_agent_cli.api._base_client.httpx.Client")
 def test_request_retries_after_rate_limited_code(mock_http_client_cls, mock_sleep):
 	auth = FakeAuthManager()
 	retrying = FakeHttpxClient(
@@ -161,9 +161,9 @@ def test_request_retries_after_rate_limited_code(mock_http_client_cls, mock_slee
 	assert mock_sleep.call_args_list[0].args[0] == 10
 
 
-@patch("boss_agent_cli.api.recruiter_client.random.uniform", return_value=0)
-@patch("boss_agent_cli.api.recruiter_client.time.sleep")
-@patch("boss_agent_cli.api.recruiter_client.httpx.Client")
+@patch("boss_agent_cli.api._base_client.random.uniform", return_value=0)
+@patch("boss_agent_cli.api._base_client.time.sleep")
+@patch("boss_agent_cli.api._base_client.httpx.Client")
 def test_request_raises_auth_error_after_max_403_retries(mock_http_client_cls, mock_sleep, mock_uniform):
 	auth = FakeAuthManager()
 	mock_http_client_cls.side_effect = [
